+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Remove unnecessary null-check in generated code for optional, disposable inputs.
+
 ### [v0.163](https://github.com/react4j/react4j/tree/v0.163) (2020-06-10) · [Full Changelog](https://github.com/react4j/react4j/compare/v0.162...v0.163) · [API Differences](https://react4j.github.io/api-diff?key=react4j&old=0.162&new=0.163)
 
 The release includes 1 non breaking API change.